% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispvar.r
\name{dispvar}
\alias{dispvar}
\title{Rank-based dispersion estimate.}
\usage{
dispvar(x, score = 1)
}
\arguments{
\item{x}{vector}

\item{score}{score type - 1 or 2}
}
\description{
This is an unbiased estimator with a correction factor for standard
deviation when normal errors.
}
\references{
T. P. Hettmansperger and J. W. McKean. Robust Nonparametric
Statistical Methods. Chapman Hall, 2012.
}
