% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getgrstplot.r
\name{getgrstplot}
\alias{getgrstplot}
\title{Q-Q Plot and Standardized Residual Plot for the GR fit.}
\usage{
getgrstplot(rlme.fit)
}
\arguments{
\item{rlme.fit}{RLME fit object}
}
\description{
It gets Q-Q Plot and Standardized Residual Plot of residuals.
}
\details{
The fit is obtained from rlme()
}
\seealso{
rlme
}
