% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getgrstplot.r
\name{getlmestplot}
\alias{getlmestplot}
\title{Q-Q Plot and Standardized Residual Plot for the REML or ML fit.}
\usage{
getlmestplot(rlme.fit)
}
\arguments{
\item{rlme.fit}{The fit is obtained from rlme()}
}
\description{
It gets Q-Q Plot and Standardized Residual Plot of residuals.
concise (1-5 lines) description of what the function does.
}
\seealso{
rlme
}
