% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{instruction}
\alias{instruction}
\title{Instruction}
\format{A data frame with 1190 observations on the following 13 variables.
\describe{ \item{list("X")}{a numeric vector} \item{list("girl")}{a numeric
vector} \item{list("minority")}{a numeric vector} \item{list("mathkind")}{a
numeric vector} \item{list("mathgain")}{a numeric vector}
\item{list("ses")}{a numeric vector} \item{list("yearstea")}{a numeric
vector} \item{list("mathknow")}{a numeric vector} \item{list("housepov")}{a
numeric vector} \item{list("mathprep")}{a numeric vector}
\item{list("classid")}{a numeric vector identifying the class within school}
\item{list("schoolid")}{a numeric vector identifying the school}
\item{list("childid")}{a numeric vector} }}
\source{
West, B., Welch, K. B., & Galecki, A. T. (2006). Linear mixed
models: a practical guide using statistical software. Chapman & Hall/CRC.
}
\description{
A data frame on school instruction results.
}
\examples{

# The following code takes a few minutes to run.
# In the interest of saving CRAN's example testing time,
# it has been commented out. If you want to use it,
# just uncomment and run.

# data(instruction)
# attach(instruction)

# data = data.frame(
#   y = mathgain,
#   mathkind = mathkind, 
#   girl = girl,
#   minority = minority,
#   ses = ses, 
#   school = factor(schoolid), 
#   section = factor(classid))


# fit.rlme = rlme(y ~ 1 + mathkind + girl + minority + ses + (1 | school) + (1 | school:section),
#  data = data,
#  method = "gr")
  
# summary(fit.rlme)

}
\keyword{datasets}

