% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/three_methods.r
\name{LM_est}
\alias{LM_est}
\title{Linear Model Estimation using the nlme package.}
\usage{
LM_est(x, y, dat, method = "REML")
}
\arguments{
\item{x}{Design matrix, (p+1)xn, with intercept.}

\item{y}{Response vector of nx1.}

\item{dat}{Data frame}

\item{method}{Character string indicating method to use, either "ML" or
"REML" (defaults to REML).}
}
\value{
\item{theta}{ Fixed effects esimates.}
  \item{ses}{ Standard error for fixed effects.}
  \item{varb}{ Variances.}
  \item{sigma}{ Error. }
  \item{ehat}{ Raw residuals }
  \item{standr.lme}{ Standardized residual }
  \item{effect_sch}{ Cluster random error. }
  \item{effect_sec}{ Subcluster random error. }
  \item{effect_err}{ Epsilon error. }
}
\description{
This gets the REML or ML estimates and predictions of random effects from
the nlme package. 
function does.
}
\references{
J. Pinheiro, D. Bates, S. DebRoy, D. Sarkar and R Development
Core Team. nlme linear and non- linear mixed effects models. The R Journal,
2011. URL http://CRAN.R-project.org/package=nlme. R package version 3.1-98.
}
\seealso{
\code{\link{rlme}}
}
\author{
Yusuf Bilgic
}
\keyword{models}
