\name{rhosch}
\alias{rhosch}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Cluster Correlation Coefficient Estimate
}
\description{
Moment estimate version of correlation coefficient in a cluster in a three-level nested design.
}
\usage{
rhosch(ahat, school, section)
}
\arguments{
  \item{ahat}{
  A vector of scores. Wilcoxon scores are used in the package.
}
  \item{school}{
  A vector of clusters.
}
  \item{section}{
  A vector of subclusters.
}
}
\references{
Y. K. Bilgic. Rank-based estimation and prediction for mixed effects models in nested designs.
2012. URL http://scholarworks.wmich.edu/dissertations/40. Dissertation.
}
