% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhoestimators.r
\name{rhosect}
\alias{rhosect}
\title{Subcluster Correlation Coefficient Estimate}
\usage{
rhosect(ahat, school, section)
}
\arguments{
\item{ahat}{A vector of scores. Wilcoxon scores are used in the package.}

\item{school}{A vector of clusters.}

\item{section}{A vector of subclusters.}
}
\description{
Moment estimate version of correlation coefficient in a subcluster in a
three-level nested design.
}
\references{
Y. K. Bilgic. Rank-based estimation and prediction for mixed
effects models in nested designs. 2012. URL
http://scholarworks.wmich.edu/dissertations/40. Dissertation.
}
