% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanresidgr.r
\name{stanresidgr}
\alias{stanresidgr}
\title{Calculate Standard Residuals}
\usage{
stanresidgr(x, y, resid, delta = 0.8, param = 2, conf = 0.95)
}
\arguments{
\item{x}{Design matrix.}

\item{y}{Response vector.}

\item{resid}{Residuals obtained from the rank-based fitting.}

\item{delta}{See HM (2012).}

\item{param}{See HM (2012).}

\item{conf}{See HM (2012).}
}
\description{
Standardizes the residuals obtained from the GR fitting.
}
\references{
T. P. Hettmansperger and J. W. McKean. Robust Nonparametric
Statistical Methods. Chapman Hall, 2012.

Y. K. Bilgic. Rank-based estimation and prediction for mixed effects models
in nested designs. 2012. URL http://scholarworks.wmich.edu/dissertations/40.
Dissertation.
}
\author{
J. W.  McKean
}
