\name{summary.rlme}
\alias{summary.rlme}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
rlme Summary
}
\description{
Summarizes a model fit from the rmle function
}
\usage{
\method{summary}{rlme}(object, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
A list of class rlme
}
  \item{\dots}{
  
}
}
\author{
Herb Susmann \email{hps1@geneseo.edu}
}

\seealso{
\code{\link{rlme}}
\code{\link{plot.rlme}}
}
