\name{wilonestep}
\alias{wilonestep}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Wilcoxon estimate for independent linear models}
\description{
This function gets weighted rank based fittings.
}
\usage{
wilonestep(y, x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{y}{
    Response vector of nx1.
}
  \item{x}{
    Design matrix, pxn, without intercept. 
}

}
\references{
J. T. Terpstra and J. W. McKean. Rank-based analysis of linear models using R. Journal of Statistical Software, 14(7) 1 -- 26, 7 2005. ISSN 1548-7660. URL http://www.jstatsoft.org/v14/i07.
}
