% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{instruction}
\alias{instruction}
\title{Instruction}
\format{A data frame with 1190 observations on the following 13 variables.
\describe{
   \item{X}{a numeric vector}
   \item{girl}{a numeric vector}
   \item{minority}{a numeric vector}
   \item{mathkind}{a numeric vector}
   \item{mathgain}{a numeric vector}
   \item{ses}{a numeric vector}
   \item{yearstea}{a numeric vector}
   \item{mathknow}{a numeric vector}
   \item{housepov}{a numeric vector}
   \item{mathprep}{a numeric vector}
   \item{classid}{a numeric vector identifying the class within school}
   \item{schoolid}{a numeric vector identifying the school}
   \item{childid}{a numeric vector}
}}
\source{
West, B., Welch, K. B., & Galecki, A. T. (2006). Linear mixed
models: a practical guide using statistical software. Chapman & Hall/CRC.
}
\description{
A data frame on school instruction results.
}
\examples{

# The following code takes a few minutes to run.
# In the interest of saving CRAN's example testing time,
# it has been commented out. If you want to use it,
# just uncomment and run.

# data(instruction)
# attach(instruction)

# data = data.frame(
#   y = mathgain,
#   mathkind = mathkind, 
#   girl = girl,
#   minority = minority,
#   ses = ses, 
#   school = factor(schoolid), 
#   section = factor(classid))


# fit.rlme = rlme(y ~ 1 + mathkind + girl + minority + ses + (1 | school) + (1 | school:section),
#  data = data,
#  method = "gr")
  
# summary(fit.rlme)
}
\keyword{datasets}
