% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlme.r
\name{plot.rlme}
\alias{plot.rlme}
\title{Plot rlme Fit}
\usage{
\method{plot}{rlme}(x, ...)
}
\arguments{
\item{x}{A list of class rlme. Store as fit.rlme.}

\item{\dots}{not used}
}
\description{
Generates Normal Q-Q plot of residuals from rlme fit
}
\examples{

data(schools)
rlme.fit = rlme(y ~ 1 + sex + age + (1 | region) + (1 | region:school), schools, method="gr")
plot(rlme.fit)

}
