% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{rlme-package}
\alias{rlme-package}
\title{rlme}
\description{
An R package for rank-based robust estimation and prediction in random
effects nested models
}
\details{
\tabular{ll}{ Package: \tab rlme\cr Type: \tab Package\cr Version: \tab
0.2\cr Date: \tab 2013-07-07\cr License: \tab GPL (>= 2)\cr }
}
\examples{


library(rlme)
data(schools)
formula = y ~ 1 + sex + age + (1 | region) + (1 | region:school)
rlme.fit = rlme(formula, schools)
summary(rlme.fit)

}
\seealso{
\code{\link{rlme}}
}
\author{
Yusuf Bilgic \email{bilgic@geneseo.edu}, Herb Susmann
\email{hps1@geneseo.edu} and Joseph McKean \email{joemckean@yahoo.com}

Maintainer: Yusuf Bilgic \email{bilgic@geneseo.edu} or
\email{yusuf.k.bilgic@gmail.com}
}
\keyword{models}
\keyword{package}
