% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{schools}
\alias{schools}
\title{PISA Literacy Data}
\format{A data frame with 334 observations on the following 6 variables.
\describe{
   \item{y}{a numeric vector indicating student literacy}
   \item{socio}{a numeric vector}
   \item{sex}{a numeric vector}
   \item{age}{a numeric vector}
   \item{region}{a numeric vector indicating four regions}
   \item{school}{a numeric vector indicating the schools within region}
}}
\description{
The data in Program for International Assessment (PISA) on academic
proficiency in schools around the world.
}
\examples{

#
# The example takes a few seconds to run, so in order to 
# save CRAN's testing time it has been commented out. 
# To run, simply uncomment and execute.
#

# data(schools)
# rlme.fit = rlme(y ~ 1 + sex + age + (1 | region) + (1 | region:school), 
#	schools, method="gr")
# summary(rlme.fit)
}
\references{
OECD (2010). PISA 2009 Results. http://www.oecd.org/
}
\keyword{datasets}
