% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlme.r
\name{summary.rlme}
\alias{summary.rlme}
\title{rlme Summary}
\usage{
\method{summary}{rlme}(object, ...)
}
\arguments{
\item{object}{A list of class rlme}

\item{\dots}{not used}
}
\description{
Summarizes a model fit from the rmle function
}
\seealso{
\code{\link{rlme}} \code{\link{plot.rlme}}
}
\author{
Herb Susmann \email{hps1@geneseo.edu}
}
