% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhoestimators.r
\name{wilonestep}
\alias{wilonestep}
\title{Wilcoxon estimate for independent linear models}
\usage{
wilonestep(y, x)
}
\arguments{
\item{y}{Response vector of nx1.}

\item{x}{Design matrix, pxn, without intercept.}
}
\description{
This function gets weighted rank based fittings.
}
\references{
J. T. Terpstra and J. W. McKean. Rank-based analysis of linear
models using R. Journal of Statistical Software, 14(7) 1 -- 26, 7 2005. ISSN
1548-7660. URL http://www.jstatsoft.org/v14/i07.
}
